+++ src/shared/utils/fileUtils.ts
@@ -1,7 +1,7 @@
 /**
- * utils.ts
+ * fileUtils.ts
  * ----------------------------------------------------------------------------
- * A collection of utility functions for file sizes and date/time handling.
+ * A collection of utility functions for file sizes handling.
  * ----------------------------------------------------------------------------
  * USAGE EXAMPLES:
  *
@@ -16,10 +16,6 @@
  * 2) Parsing a file size:
  *    const bytes = parseFileSize("2 MB");
  *    // => 2097152
- *
- * 3) Formatting a date/time:
- *    const dateStr = formatDateTime(new Date(), { includeTime: true, locale: 'en-GB' });
- *    // => e.g. "12 September 2025 10:30"
  */
 
 // ----------------------------------------------------------------------------
@@ -42,6 +38,8 @@ export interface FormatFileSizeOptions {
 }
 
 /**
+ * formatFileSize
+ * ----------------------------------------------------------------------------
  * Convert a numeric byte size into a human-readable string (e.g. "12.34 MB").
  *
  * @param bytes     The file size in bytes.
@@ -77,6 +75,8 @@ export function formatFileSize(
 }
 
 /**
+ * parseFileSize
+ * ----------------------------------------------------------------------------
  * Parses a string like "20 KB" or "5.3 MB" into a numeric size in bytes.
  *
  * @param sizeString   The file size string, e.g. "2 MB", "1.5 GB", "4096 Bytes"
@@ -113,78 +113,4 @@ export function parseFileSize(sizeString: string, throwOnError = true): number {
 			if (throwOnError) throw new Error(`Unknown unit: ${unit}`);
 			return NaN;
 	}
-}
-
-// ----------------------------------------------------------------------------
-// DATE/TIME UTILITIES
-// ----------------------------------------------------------------------------
-
-export interface FormatDateTimeOptions {
-	/**
-	 * Whether to include time (HH:MM) in the output. Defaults to false.
-	 */
-	includeTime?: boolean;
-	/**
-	 * The locale for date/time formatting (e.g. 'en-US', 'fr-FR'). Defaults to 'en-US'.
-	 */
-	locale?: string;
-	/**
-	 * A time zone identifier (e.g. 'UTC', 'America/Los_Angeles'). If omitted,
-	 * it uses the system default.
-	 */
-	timeZone?: string;
-}
-
-/**
- * Format a Date or date string into a human-readable format.
- *
- * @param date    The date object or string to format.
- * @param options Optional formatting config, e.g. { includeTime: true, locale: 'en-GB', timeZone: 'UTC' }
- * @returns A formatted date/time string, locale-aware.
- */
-export function formatDateTime(
-	date: string | Date,
-	{ includeTime, locale = 'en-US', timeZone }: FormatDateTimeOptions = {},
-): string {
-	const parsedDate = typeof date === 'string' ? new Date(date) : date;
-
-	const dateFormat: Intl.DateTimeFormatOptions = {
-		year: 'numeric',
-		month: 'long',
-		day: 'numeric',
-		...(timeZone ? { timeZone } : {}),
-	};
-
-	const timeFormat: Intl.DateTimeFormatOptions = {
-		hour: 'numeric',
-		minute: 'numeric',
-		hour12: true,
-		...(timeZone ? { timeZone } : {}),
-	};
-
-	// Build up the date string
-	let datePart = new Intl.DateTimeFormat(locale, dateFormat).format(parsedDate);
-
-	if (includeTime) {
-		const timePart = new Intl.DateTimeFormat(locale, timeFormat).format(parsedDate);
-		datePart += ` ${timePart}`;
-	}
-
-	return datePart;
-}
-
-/**
- * computeExpirationDays
- * ----------------------------------------------------------------------------
- * Calculates how many whole days between `expirationTime` and now.
- *
- * @param expirationTime - A date string in ISO format (e.g., "2025-12-01T00:00:00Z")
- * @returns Number of days (rounded up)
- */
-export function computeExpirationDays(expirationTime: string): number {
-	if (!expirationTime) return 0;
-	const expirationDate = new Date(expirationTime);
-	const now = new Date();
-	const diffTime = Math.abs(expirationDate.getTime() - now.getTime());
-	return Math.ceil(diffTime / (1000 * 60 * 60 * 24));
 }
